chore: apply pot to WebEmbedded instead of Web
