+++ common/src/main/java/dev/lavalink/youtube/clients/skeleton/NonMusicClient.java
@@ -75,11 +75,8 @@ public abstract class NonMusicClient implements Client {
 
         ClientConfig config = getBaseClientConfig(httpInterface);
 
-        if (config.getName() == "WEB_EMBEDDED_PLAYER" || status == null) {
-            // Only add embed info if the status is not NON_EMBEDDABLE.
-            config.withClientField("clientScreen", "EMBED")
-                .withThirdPartyEmbedUrl("https://google.com");
-        }
+        config.withClientField("clientScreen", "EMBED")
+            .withThirdPartyEmbedUrl("https://google.com");
 
         String payload = config.withRootField("videoId", videoId)
             .withRootField("racyCheckOk", true)
