+++ plugin/src/main/java/dev/lavalink/youtube/plugin/YoutubeRestHandler.java
@@ -3,7 +3,7 @@ package dev.lavalink.youtube.plugin;
 import com.sedmelluq.discord.lavaplayer.player.AudioPlayerManager;
 import com.sedmelluq.discord.lavaplayer.tools.DataFormatTools;
 import dev.lavalink.youtube.YoutubeAudioSourceManager;
-import dev.lavalink.youtube.clients.Web;
+import dev.lavalink.youtube.clients.WebEmbedded;
 import dev.lavalink.youtube.plugin.rest.MinimalConfigRequest;
 import dev.lavalink.youtube.plugin.rest.MinimalConfigResponse;
 import org.slf4j.Logger;
@@ -55,7 +55,7 @@ public class YoutubeRestHandler {
         String visitorData = config.getVisitorData();
 
         if (poToken == null || visitorData == null || (!poToken.isEmpty() && !visitorData.isEmpty())) {
-            Web.setPoTokenAndVisitorData(poToken, visitorData);
+            WebEmbedded.setPoTokenAndVisitorData(poToken, visitorData);
             log.debug("Updated poToken to \"{}\" and visitorData to \"{}\"", poToken, visitorData);
         }
     }
